+++ CHANGELOG.markdown
@@ -2,6 +2,10 @@
 
 ## unreleased
 
+* Use thread pool instead of creating a thread for every iteration[@toy](https://github.com/toy)
+* Handle `break` (also with argument) and exceptions raised in `each` method of enumerable [@toy](https://github.com/toy)
+* Register `lazy` to run without threads [@toy](https://github.com/toy)
+
 ## v1.4.0 (2017-03-19)
 
 * Register `sum` and `uniq` to run in threads and `chunk_while` to run without threads [@toy](https://github.com/toy)